+++ lib/questions.js
@@ -3,6 +3,7 @@ import _ from 'lodash';
 import { Dice1, Dice2, Dice3, Dice4, Dice5, Dice6 } from 'react-bootstrap-icons';
 import ChessBoard from '../components/chessboard';
 import { setQuestionsTv } from './questions-tv';
+import { setQuestionsPatterns } from './questions-patterns';
 
 export const demoIds = {
     ADD_1: 'add-1',
@@ -248,37 +249,7 @@ export function generateDemoQuestion({ demoId, subject }) {
 
             case demoIds.PATTERNS_1: {
 
-                // Question is a pattern of 6 circle emojis.
-                // Available patterns are:
-                // - AABBCC
-                // - ABABAB
-                // - ABCABC
-                // - AAABBB
-                
-                // The colors are randomly selected from the list of colors.
-                const colors = ['🔴', '🟠', '🟡', '🟢', '🔵', '🟣', '🟤', '⚫', '⚪'];
-                let color = {};
-                color.A = _.sample(colors);
-                color.B = _.sample(_.without(colors, color.A));
-                color.C = _.sample(_.without(colors, color.A, color.B));
-                
-                // The pattern is randomly selected from the list of patterns.
-                const patterns = ['AABBCC', 'ABABAB', 'ABCABC', 'AAABBB'];
-                const pattern = _.sample(patterns);
-                const position = _.random(0, 5);
-                
-                text = _.range(6).map((i) => {
-                    if (i === position) return '❓';
-                    return color[pattern[i]];
-                }).join('');
-                solution = color[pattern[position]];
-
-                // The options are the colors that are not in the pattern.
-                options = _.shuffle([
-                    { id: 1, value: color.A },
-                    { id: 2, value: color.B },
-                    { id: 3, value: color.C },
-                ]);
+                ({ text, solution, options } = setQuestionsPatterns(text, solution, options));
 
                 break;
             }
